requirements.txt
fastapi
uvicorn
python-multipart
paddleocr
paddlepaddle
opencv-python-headless
pandas
openpyxl
pydantic
aiofiles
